move auth routes under users router. add logout option to navbar after succesful login
