+++ views/layout.pug
@@ -22,6 +22,9 @@ html(lang='en' class='h-100')
             li(class='nav-item active')
               a(class='nav-link' href='/') Home
                 span(class='sr-only') (current)
+            if user
+              li(class='nav-item active')
+                a(class='nav-link' href='/users/logout') Logout            
     main(role='main' class='flex-shrink-0')
       div(class='container')
